new short Array template in java_lang.cc

+++ tests/TestShortArray.java
@@ -5,7 +5,7 @@ public class TestShortArray {
   
   public static void main ( String[] args ) {
   
-    int[] a0 = new int[0];
+    short[] a0 = new short[0];
 
     if (a0.length == 0) {
       System.out.println("PASS short[0].length == 0");
